2021.8.5  17:57
将受到伤害的逻辑封装到Unit对象中
增加建筑受伤展示
增加收到生命恢复的展示
实现建筑维修站的功能
修改武当太极剑客的特效
峨眉派建筑、科技、兵种加入游戏

+++ Assets/Scripts/GameInfo.cs
@@ -188,8 +188,8 @@ public class Projectile
         {
             attack_type = Convert.ToInt32(info[10]);
             attack_mood = Convert.ToInt32(info[11]);
-            speed = Convert.ToInt32(info[12]);
-            distance = Convert.ToInt32(info[13]);
+            speed = (float)Convert.ToDouble(info[12]);
+            distance = (float)Convert.ToDouble(info[13]);
             lockobj = Convert.ToInt32(info[14]);
             value = Convert.ToInt32(info[15]);
             destroy = Convert.ToInt32(info[16]);
